fib program

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,13 +9,13 @@
  */
 int main(void)
 {
-	int i, x = 1, y = 2, z;
+	long long i, x = 1, y = 2, z;
 
-	printf("%d, %d, ", x, y);
+	printf("%lld, %lld, ", x, y);
 	for (i = 0; i < 48; i++)
 	{
 		z = x + y;
-		printf("%d", z);
+		printf("%lld", z);
 		if (i < 47)
 			printf(", ");
 		x = y;
